backend/pyproject.toml
[tool.poetry]
name = "english-fight-api"
version = "0.1.0"
description = "Backend api for english fight VK app"
authors = ["Dima Krutolianov <DimaKrutDK@ya.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.15"
django-environ = "^0.4.5"
mysqlclient = "^2.0.1"
djangorestframework = "^3.12.2"
drf-flex-fields = "^0.8.6"
django-cors-headers = "^3.5.0"
click = "^7.1.2"
django-click = "^2.2.0"
django-lifecycle = "^0.8.0"
django-cprofile-middleware = "^1.0.5"
sentry-sdk = "^0.19.5"
django-admin-honeypot = "^1.1.0"
vk-api = "^11.9.1"
typing-extensions = "^3.7.4"
django-hashid-field = "^3.1.3"
channels = "^3.0.3"
channels-redis = "^3.2.0"
daphne = "^3.0.1"
whitenoise = "^6.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-django = "^4.1.0"
snakeviz = "^2.1.0"
black = "^20.8b1"
flake8 = "^3.8.4"
bandit = "^1.7.0"
pytest-asyncio = "^0.14.0"

[tool.black]
line-length = 79
exclude="migrations"
skip-string-normalization=true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
